refactor: moved Floor related qurries to own file
