+++ IndoorMapsFrontend/src/schema.graphql
@@ -3,6 +3,33 @@
 # !!!   DO NOT MODIFY THIS FILE BY YOURSELF   !!!
 # -----------------------------------------------
 
+type Area {
+  category: String!
+  databaseId: Int!
+  description: String!
+  id: ID!
+  shape: String!
+  title: String!
+  traversable: Boolean!
+}
+
+input AreaCreateInput {
+  category: String
+  description: String!
+  floorDatabseId: Int!
+  shape: String!
+  title: String!
+}
+
+input AreaModifyInput {
+  category: String
+  description: String
+  floorDatabseId: Int
+  id: Int!
+  shape: String
+  title: String
+}
+
 type Building {
   address: String!
   databaseId: Int!
@@ -36,6 +63,7 @@ enum EditorLevel {
 }
 
 type Floor {
+  areas: [Area!]!
   databaseId: Int!
   description: String!
   id: ID!
@@ -76,12 +104,20 @@ type LogedInUser {
 }
 
 type Mutation {
+  createArea(data: AreaCreateInput!): NewAreaResult!
   createBuilding(data: BuildingCreateInput!): Building!
   createFloor(data: FloorCreateInput!): NewFloorResult!
+  modifyArea(data: AreaModifyInput!): NewAreaResult!
   modifyFloor(data: FloorModifyInput!): NewFloorResult!
   signOut: SignedOutSuccess!
   signinUser(data: UserLoginInput!): User!
   signupUser(data: UserCreateInput!): User!
+}
+
+type NewAreaResult {
+  databaseId: Int!
+  floorDatabaseId: Int!
+  success: Boolean!
 }
 
 type NewFloorResult {
